SM-12 fix: moved components from App.tsx to separate files
