+++ src/features/adminPage/AdminPage.tsx
@@ -1,5 +1,18 @@
-const AdminPage = () => {
-  return <div>AdminPage</div>;
+const mainStyle = {
+  background: 'grey',
+  display: 'flex',
+  margin: '0 auto',
+  width: '80%',
+  height: 'calc(100vh - 4rem)',
+  justifyContent: 'space-between',
+  padding: '1rem 1rem',
 };
 
+const AdminPage = () => {
+  return (
+    <div className="text-3xl font-bold underline" style={mainStyle}>
+      AdminPage
+    </div>
+  );
+};
 export default AdminPage;
