+++ src/features/homePage/HomePage.tsx
@@ -1,5 +1,19 @@
+const mainStyle = {
+  background: 'grey',
+  display: 'flex',
+  margin: '0 auto',
+  width: '80%',
+  height: 'calc(100vh - 4rem)',
+  justifyContent: 'space-between',
+  padding: '1rem 1rem',
+};
+
 const HomePage = () => {
-  return <div>HomePage</div>;
+  return (
+    <div className="text-3xl font-bold underline" style={mainStyle}>
+      HomePage
+    </div>
+  );
 };
 
 export default HomePage;
