--- src/features/layout/cart/Cart.tsx
@@ -1,5 +1,0 @@
-const Cart = () => {
-  return <div>Cart</div>;
-};
-
-export default Cart;
