+++ src/features/products/productSeries/ProductSeriesPage.tsx
@@ -1,5 +1,18 @@
-const ProductSeriesPage = () => {
-  return <div>ProductSeriesPage</div>;
+const mainStyle = {
+  background: 'grey',
+  display: 'flex',
+  margin: '0 auto',
+  width: '80%',
+  height: 'calc(100vh - 4rem)',
+  justifyContent: 'space-between',
+  padding: '1rem 1rem',
 };
 
+const ProductSeriesPage = () => {
+  return (
+    <div className="text-3xl font-bold underline" style={mainStyle}>
+      ProductSeriesPage
+    </div>
+  );
+};
 export default ProductSeriesPage;
